Fix popovers across site
